refactor: move Rust solver to crates/solver/ with dual WASM/PyO3 feature flags
Move the FISTA solver from wasm/catune-solver/ to crates/solver/ (standard
Rust convention) and rename the crate from catune-solver to calab-solver to
match the monorepo identity. This prepares the crate for multi-target builds.

Key changes:
- Make wasm-bindgen conditional behind a "jsbindings" feature flag (default)
- Add PyO3 stub behind a "pybindings" feature flag (compiles, not yet functional)
- Pure library mode (no features) compiles for use as a Rust dependency
- Rebuild WASM package with new calab_solver naming
- Update all path references across build scripts, CI, ESLint, docs (18 files)

The WASM build and all 29 Rust tests remain fully operational.
