+++ crates/solver/.cargo/config.toml
