+++ crates/solver/pkg/calab_solver.d.ts
