+++ crates/solver/pkg/calab_solver_bg.wasm.d.ts
