+++ crates/solver/src/fft.rs
