+++ crates/solver/src/filter.rs
