+++ crates/solver/src/kernel.rs
