+++ docs/CONTRIBUTING.md
@@ -84,7 +84,7 @@ Run `npm run lint && npm run format:check && npm run typecheck` before pushing.
 
 ESLint enforces these import boundaries:
 
-- **WASM**: Only `packages/core/src/wasm-adapter.ts` may import from `wasm/catune-solver/pkg/`
+- **WASM**: Only `packages/core/src/wasm-adapter.ts` may import from `crates/solver/pkg/`
 - **Supabase**: Only `packages/community/src/supabase.ts` may import `@supabase/supabase-js`
 - **Package barrels**: App files import from `@calab/<pkg>`, never from `@calab/<pkg>/src/*`
