+++ python/tests/test_cross_language.py
@@ -3,8 +3,8 @@
 Verifies that the Python solver produces results matching the Rust solver
 within numerical tolerance (f32 vs f64 divergence over iterations).
 
-Fixtures are generated by `wasm/catune-solver/tests/generate_fixtures.rs`.
-Run `cd wasm/catune-solver && cargo test generate_fixtures -- --ignored`
+Fixtures are generated by `crates/solver/tests/generate_fixtures.rs`.
+Run `cd crates/solver && cargo test generate_fixtures -- --ignored`
 to regenerate.
 """
