--- wasm/catune-solver/pkg/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "catune-solver",
-  "type": "module",
-  "version": "0.1.0",
-  "files": [
-    "catune_solver_bg.wasm",
-    "catune_solver.js",
-    "catune_solver.d.ts"
-  ],
-  "main": "catune_solver.js",
-  "types": "catune_solver.d.ts",
-  "sideEffects": [
-    "./snippets/*"
-  ]
-}
